added long w crashes inserted -- not pre/post yet

+++ inputs/long_w_single_nid.rs
@@ -1,31 +1,31 @@
 #[atomic]
 #[nids(x)]
-fn update1(x: &mut u16) {
+fn update1(x: &mut i32) {
     *x = 10;
 }
 
 #[atomic]
 #[nids(x)]
-fn update2(x: &mut u16) {
+fn update2(x: &mut i32) {
     *x = 20;
     *x = 25;
 }
 
 #[atomic]
 #[nids(x)]
-fn update3(x: &mut u16) {
+fn update3(x: &mut i32) {
     *x = *x + 5;
 }
 
 #[atomic]
 #[nids(x)]
-fn update4(x: &mut u16) {
+fn update4(x: &mut i32) {
     *x = 0;
 }
 
 #[atomic]
 #[nids(x)]
-fn update5(x: &mut u16) {
+fn update5(x: &mut i32) {
     *x = 100;
     *x = 50;
     *x = 100;
@@ -33,19 +33,19 @@ fn update5(x: &mut u16) {
 
 #[atomic]
 #[nids(x)]
-fn update6(x: &mut u16) {
+fn update6(x: &mut i32) {
     *x = 0xFFFF;
 }
 
 #[atomic]
 #[nids(x)]
-fn update7(x: &mut u16) {
+fn update7(x: &mut i32) {
     *x = 123;
 }
 
 #[atomic]
 #[nids(x)]
-fn update8(x: &mut u16) {
+fn update8(x: &mut i32) {
     *x = 8;
     *x = 3;
     *x = 8;
@@ -54,7 +54,7 @@ fn update8(x: &mut u16) {
 #[atomic]
 #[nids(x)]
 fn main() {
-    let mut x: u16 = 0;
+    let mut x: i32 = 0;
 
     update1(&mut x);
     update2(&mut x);
